+++ src/config/chain/sei.ts
@@ -1,14 +1,23 @@
+import { wallets as compassWallets } from "@cosmos-kit/compass";
+import { wallets as keplrWallets } from "@cosmos-kit/keplr";
+
 import type { ChainConfigs } from "./types";
 
+/**
+ * @remarks
+ * Because Sei utilizes Cosmos SDK v0.26, which lacks the anyOfAddresses feature, we need to disable it.
+ * We already have disabled only normal upload section, not proposal to store code section.
+ */
 export const SEI_CHAIN_CONFIGS: ChainConfigs = {
   "pacific-1": {
     chain: "sei",
     registryChainName: "sei",
     prettyName: "Sei",
     lcd: "https://sei-api.polkachu.com",
-    rpc: "https://sei-rpc.polkachu.com",
+    rpc: "https://sei-rpc.polkachu.com:443",
     indexer: "https://pacific-1-graphql.alleslabs.dev/v1/graphql",
     api: "https://celatone-api.alleslabs.dev",
+    wallets: [...compassWallets, ...keplrWallets],
     features: {
       faucet: {
         enabled: false,
@@ -23,6 +32,9 @@ export const SEI_CHAIN_CONFIGS: ChainConfigs = {
       },
       publicProject: {
         enabled: true,
+      },
+      gov: {
+        enabled: false,
       },
     },
     gas: {
@@ -37,15 +49,19 @@ export const SEI_CHAIN_CONFIGS: ChainConfigs = {
       validator: "",
       proposal: "",
     },
+    extra: {
+      disableAnyOfAddresses: true,
+    },
   },
   "atlantic-2": {
     chain: "sei",
     registryChainName: "seitestnet2",
     prettyName: "Sei Testnet2",
     lcd: "https://rest.atlantic-2.seinetwork.io",
-    rpc: "https://rpc.atlantic-2.seinetwork.io",
+    rpc: "https://rpc.atlantic-2.seinetwork.io:443",
     indexer: "https://atlantic-2-graphql.alleslabs.dev/v1/graphql",
     api: "https://celatone-api.alleslabs.dev",
+    wallets: [...compassWallets, ...keplrWallets],
     features: {
       faucet: {
         enabled: false,
@@ -61,6 +77,9 @@ export const SEI_CHAIN_CONFIGS: ChainConfigs = {
       publicProject: {
         enabled: false,
       },
+      gov: {
+        enabled: false,
+      },
     },
     gas: {
       gasPrice: {
@@ -73,6 +92,9 @@ export const SEI_CHAIN_CONFIGS: ChainConfigs = {
     explorerLink: {
       validator: "https://testnet.mintscan.io/sei-testnet/validators",
       proposal: "https://testnet.mintscan.io/sei-testnet/proposals",
+    },
+    extra: {
+      disableAnyOfAddresses: true,
     },
   },
 };