+++ src/config/chain/stargaze.ts
@@ -1,3 +1,5 @@
+import { wallets as keplrWallets } from "@cosmos-kit/keplr";
+
 import type { ChainConfigs } from "./types";
 
 export const STARGAZE_CHAIN_CONFIGS: ChainConfigs = {
@@ -6,9 +8,10 @@ export const STARGAZE_CHAIN_CONFIGS: ChainConfigs = {
     registryChainName: "stargaze",
     prettyName: "Stargaze",
     lcd: "https://rest.stargaze-apis.com",
-    rpc: "https://rpc.stargaze-apis.com",
+    rpc: "https://rpc.stargaze-apis.com:443",
     indexer: "https://stargaze-mainnet-graphql.alleslabs.dev/v1/graphql",
     api: "https://celatone-api.alleslabs.dev",
+    wallets: [...keplrWallets],
     features: {
       faucet: {
         enabled: false,
@@ -24,6 +27,9 @@ export const STARGAZE_CHAIN_CONFIGS: ChainConfigs = {
       publicProject: {
         enabled: true,
       },
+      gov: {
+        enabled: true,
+      },
     },
     gas: {
       gasPrice: {
@@ -37,5 +43,48 @@ export const STARGAZE_CHAIN_CONFIGS: ChainConfigs = {
       validator: "https://www.mintscan.io/stargaze/validators",
       proposal: "https://www.mintscan.io/stargaze/proposals",
     },
+    extra: {},
+  },
+  "elgafar-1": {
+    chain: "stargaze",
+    registryChainName: "stargazetestnet",
+    prettyName: "Stargaze Testnet",
+    lcd: "https://rest.elgafar-1.stargaze-apis.com",
+    rpc: "https://rpc.elgafar-1.stargaze-apis.com",
+    indexer: "https://elgafar-1-graphql.alleslabs.dev/v1/graphql",
+    api: "https://celatone-api.alleslabs.dev",
+    wallets: [...keplrWallets],
+    features: {
+      faucet: {
+        enabled: false,
+      },
+      wasm: {
+        enabled: true,
+        storeCodeMaxFileSize: 800_000,
+        clearAdminGas: 50_000,
+      },
+      pool: {
+        enabled: false,
+      },
+      publicProject: {
+        enabled: true,
+      },
+      gov: {
+        enabled: true,
+      },
+    },
+    gas: {
+      gasPrice: {
+        tokenPerGas: 0.04,
+        denom: "ustars",
+      },
+      gasAdjustment: 1.5,
+      maxGasLimit: 50_000_000,
+    },
+    explorerLink: {
+      validator: "https://testnet-explorer.publicawesome.dev/stargaze/staking",
+      proposal: "https://testnet-explorer.publicawesome.dev/stargaze/gov",
+    },
+    extra: {},
   },
 };