+++ src/lib/app-provider/contexts/app.tsx
@@ -11,13 +11,12 @@ import {
   createContext,
 } from "react";
 
-import { useAmplitude } from "../hooks/useAmplitude";
 import { useNetworkChange } from "../hooks/useNetworkChange";
 import { CHAIN_CONFIGS, DEFAULT_CHAIN_CONFIG } from "config/chain";
 import type { ChainConfig } from "config/chain";
 import { PROJECT_CONSTANTS } from "config/project";
 import type { ProjectConstants } from "config/project";
-import { SUPPORTED_CHAIN_IDS } from "env";
+import { HASURA_ADMIN_SECRET, SUPPORTED_CHAIN_IDS } from "env";
 import { LoadingOverlay } from "lib/components/LoadingOverlay";
 import { NetworkErrorState } from "lib/components/state/NetworkErrorState";
 import { DEFAULT_ADDRESS } from "lib/data";
@@ -71,7 +70,11 @@ export const AppProvider = observer(({ children }: AppProviderProps) => {
       availableChainIds: SUPPORTED_CHAIN_IDS,
       currentChainId,
       chainConfig,
-      indexerGraphClient: new GraphQLClient(chainConfig.indexer),
+      indexerGraphClient: new GraphQLClient(chainConfig.indexer, {
+        headers: {
+          "x-hasura-admin-secret": HASURA_ADMIN_SECRET,
+        },
+      }),
       constants: PROJECT_CONSTANTS,
     };
   }, [currentChainId]);
@@ -102,8 +105,6 @@ export const AppProvider = observer(({ children }: AppProviderProps) => {
   }, [setModalTheme]);
 
   useNetworkChange(handleOnChainIdChange);
-
-  useAmplitude(currentChainName);
 
   if (currentChainId && !(currentChainId in CHAIN_CONFIGS))
     return <NetworkErrorState />;