+++ src/lib/app-provider/tx/upload.ts
@@ -20,7 +20,7 @@ export type UploadSucceedCallback = (txResult: UploadTxInternalResult) => void;
 export interface UploadStreamParams {
   wasmFileName: Option<string>;
   wasmCode: Option<Promise<ArrayBuffer>>;
-  addresses: Addr[];
+  addresses?: Addr[];
   permission: AccessType;
   codeName: string;
   estimatedFee: Option<StdFee>;