+++ src/lib/providers/query-client.tsx
@@ -1,0 +1,11 @@
+import {
+  QueryClient,
+  QueryClientProvider as Provider,
+} from "@tanstack/react-query";
+import type { ReactNode } from "react";
+
+export const QueryClientProvider = ({ children }: { children: ReactNode }) => {
+  const queryClient = new QueryClient();
+
+  return <Provider client={queryClient}>{children}</Provider>;
+};